requirements.txt
# PyTorch with M1 Mac support
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Core dependencies
numpy==1.24.3
pillow==9.5.0
matplotlib==3.7.1
scikit-learn==1.2.2
PyYAML==6.0
tqdm==4.65.0
seaborn==0.12.2
pandas==2.0.2
opencv-python==4.7.1.72

# M1 Mac Installation Notes:
# Install with: pip install -r requirements.txt
# For M1 Macs, PyTorch should automatically detect MPS support
# If issues occur, try: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
